requirements.txt
slackeventsapi~=2.2.1
Flask~=1.1.4
python-dotenv~=0.17.1
psycopg2~=2.8.6
pandas~=1.2.4
slackclient~=2.9.3
